# docs/api-documentation.yaml
openapi: 3.0.0
info:
  title: Course Management Platform API
  version: 1.0.0
  description: Comprehensive API for course management with multi-role support
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.courseplatform.com/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /forum/threads:
    get:
      summary: Get forum threads
      tags: [Forum]
      parameters:
        - in: query
          name: courseId
          schema:
            type: string
        - in: query
          name: categoryId
          schema:
            type: string
      responses:
        200:
          description: List of threads

    post:
      summary: Create thread
      tags: [Forum]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                courseId:
                  type: string
                categoryId:
                  type: string
      responses:
        201:
          description: Thread created
  /forum/threads/{id}/replies:
    post:
      summary: Reply to thread
      tags: [Forum]
      security:
        - bearerAuth: []
      responses:
        201:
          description: Reply added
  /forum/threads/{id}/like:
    post:
      summary: Like thread
      tags: [Forum]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Thread liked

  # ==========================================
  # ANNOUNCEMENTS
  # ==========================================
...